초기 로드 시 gridView 뜨도록 수정 및 구분을 위해 기존 pressNews.js -> gridView.js로 수정
