+++ index.html
@@ -42,6 +42,6 @@
         <script src="src/components/fieldTab.js"></script>
         <script type="module" src="./src/index.js"></script> 
         <script type="module" src="src/components/newsTabs.js"></script>
-        <script type="module" src="src/components/pressNews.js"></script>
+        <script type="module" src="src/components/gridView.js"></script>
     </body>
 </html>
