+++ src/components/newsTabs.js
@@ -1,4 +1,4 @@
-
+import { showAllNewsBrands } from "./gridView.js";
 const allNewsArticles = [
     "전체 언론사 기사 1",
     "전체 언론사 기사 2",
@@ -23,8 +23,23 @@ function updateGrid(articles) {
         gridView.appendChild(articleDiv);
     });
     console.log('gridView:', gridView);
+    
+    showAllNewsBrands();
+}
+
+function updateMyGrid(articles) {
+    const gridView = document.getElementById('gridView');
+    gridView.innerHTML = '';
+
+    articles.forEach(article => {
+        const articleDiv = document.createElement('div');
+        articleDiv.textContent = article;
+        articleDiv.classList.add('article');
+        gridView.appendChild(articleDiv);
+    });
+    console.log('gridView:', gridView);
     if (gridView) {
-        gridView.innerHTML = "언론사 표시하기!!";
+        gridView.innerHTML = "구독 중인 언론사!!";
     } else {
         console.error("gridView 요소를 찾을 수 없습니다.");
     }
@@ -49,7 +64,7 @@ document.getElementById('allNewsBtn').addEventListener('click', () => {
 });
 
 document.getElementById('subNewsBtn').addEventListener('click', () => {
-    updateGrid(subscribedNewsArticles); // 구독 언론사
+    updateMyGrid(subscribedNewsArticles); // 구독 언론사
     updateButtonState('subNewsBtn');
 });
 
@@ -57,4 +72,5 @@ document.getElementById('subNewsBtn').addEventListener('click', () => {
 window.onload = () => {
     updateGrid(allNewsArticles); // 초기 로드 시 전체 언론사 콘텐츠 표시
     updateButtonState('allNewsBtn');
+    showAllNewsBrands();
 };
